% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b_1_iv.R
\name{B.1.IV}
\alias{B.1.IV}
\title{Beta value calculation for simple IIV}
\usage{
B.1.IV(dep.var, endo.regressor, instrument)
}
\arguments{
\item{dep.var}{the dependend variable (y)}

\item{endo.regressor}{the regressor that is correlated with the error term(x)}

\item{instrument}{the (imperfect) instrument used to estimate the effect of the endogenous regressor}
}
\value{
the coefficient value for single iiv estimation
}
\description{
This function calculates the beta as seen in the estimation equations
}
\examples{
B.1.IV()
}
\keyword{beta}
\keyword{bound}
\keyword{iiv,}
\keyword{value,}
