% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_direction.R
\name{Multiple.Direction}
\alias{Multiple.Direction}
\title{Beta value calculation for multiple iv regressions}
\usage{
Multiple.Direction(X.wiggle, endo.regressor, instrument)
}
\arguments{
\item{X.wiggle}{the regressor matrix with the variance of the exogenous regressors filtered out}

\item{endo.regressor}{the regressor that is correlated with the error term(x)}

\item{instrument}{the (imperfect) instrument used to estimate the effect of the endogenous regressor}
}
\value{
the criteria needed for multiple iiv regression
}
\description{
This function calculates the criterion as seen in the paper.
}
\examples{
Multiple.Direction()
}
\keyword{beta}
\keyword{bound}
\keyword{iiv,}
\keyword{value,}
