% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_linear_iiv_coef.R
\name{Multiple.Linear.IIV.beta}
\alias{Multiple.Linear.IIV.beta}
\title{Imperfect instrumental variable with one or more imperfect instruments.}
\usage{
Multiple.Linear.IIV.beta(data, dep.var, endo.regressor, instrument,
  exog.regressors = c(), valid.instruments = c(),
  covar.direction = "xxxx")
}
\arguments{
\item{data}{a dataframe that contains all data}

\item{dep.var}{the dependend variable (y)}

\item{endo.regressor}{the regressor that is correlated with the error term(x)}

\item{instrument}{the (imperfect) instrument used to estimate the effect of the endogenous regressor}

\item{exog.regressors}{the regressors that are not correlated with dep.var and the error term}

\item{valid.instruments}{instruments that can be considered independent of the error term}

\item{covar.direction}{(string) direction of the covariance}
}
\value{
beta estimator as well as indication if it is a one or two-sided bound
}
\description{
This estimator calculates the bounds of the ivv coefficients when one ore more regressors are exogenous
and/or can be estimated with valid instruments
}
\examples{
Multiple.Linear.IIV.beta()
}
\keyword{endogenous}
\keyword{iiv,}
\keyword{one}
\keyword{regressor}
\keyword{simple,}
