% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_coefs_iiv.R
\name{Multiple.Linear.IIV.other}
\alias{Multiple.Linear.IIV.other}
\title{Other coefficients for a multiple imperfect instrumental variable regression}
\usage{
Multiple.Linear.IIV.other(data, dep.var, endo.regressor,
  exog.regressors = c(), valid.instruments = c(), beta.low, beta.upp)
}
\arguments{
\item{data}{a dataframe that contains all data}

\item{dep.var}{the dependend variable (y)}

\item{endo.regressor}{the regressor that is correlated with the error term(x)}

\item{exog.regressors}{the regressors that are not correlated with dep.var and the error term}

\item{valid.instruments}{instruments that can be considered independent of the error term}

\item{beta.low}{the lower bound of the estimation for the endogenous variable}

\item{beta.upp}{the upper bound of the estimation for the endogenous variable}
}
\value{
estimators and std. errors for the exogenous variables in the iiv equation
}
\description{
This estimator calculates the bounds and standard errors for the other, non endogenous variables in the 
iiv model.
}
\examples{
Multiple.Linear.IIV.other()
}
\keyword{endogenous}
\keyword{iiv,}
\keyword{one}
\keyword{regressor}
\keyword{simple,}
