% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_linear_iiv_a4.R
\name{Simple.Linear.IIV.A4}
\alias{Simple.Linear.IIV.A4}
\title{Imperfect instrumental variable estimator with one imperfect instrument. Additionally the assumption 4
from the paper is imposed (the instrument is less endogenous than the endogenous regressor)}
\usage{
Simple.Linear.IIV.A4(data, dep.var, endo.regressor, instrument,
  cor.direction = "xxxx")
}
\arguments{
\item{data}{a dataframe that contains all data}

\item{dep.var}{the dependend variable (y)}

\item{endo.regressor}{the regressor that is correlated with the error term(x)}

\item{instrument}{the (imperfect) instrument used to estimate the effect of the endogenous regressor}

\item{cor.direction}{(string) direction of the correlation}
}
\value{
the estimator bounds as well as an indication about the nature of the bounds under a4
}
\description{
This estimator calculates the bounds of an IIV with one endogenous regressor, with A4 as well
}
\examples{
Simple.Linear.IIV.A4()
}
\keyword{endogenous}
\keyword{iiv,}
\keyword{instrument}
\keyword{less}
\keyword{one}
\keyword{regressor,}
\keyword{simple,}
