% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_linear_iiv.R
\name{Simple.Linear.IIV}
\alias{Simple.Linear.IIV}
\title{Imperfect instrumental variable estimator with one imperfect instrument}
\usage{
Simple.Linear.IIV(data, dep.var, endo.regressor, instrument,
  cor.direction = "xxxx")
}
\arguments{
\item{data}{a dataframe that contains all data}

\item{dep.var}{the dependend variable (y)}

\item{endo.regressor}{the regressor that is correlated with the error term(x)}

\item{instrument}{the (imperfect) instrument used to estimate the effect of the endogenous regressor}

\item{cor.direction}{(string) direction of the correlation}
}
\value{
the estimator bounds as well as an indication about the nature of the bounds
}
\description{
This estimator calculates the bounds of an IIV with one endogenous regressor
}
\examples{
Simple.Linear.IIV()
}
\keyword{endogenous}
\keyword{iiv,}
\keyword{one}
\keyword{regressor}
\keyword{simple,}
