% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two_instr_bound.R
\name{Two.Instr.Bounds}
\alias{Two.Instr.Bounds}
\title{Imperfect instrumental variable estimation when more than one imperfect instrument is available}
\usage{
Two.Instr.Bounds(data, endo.regressor, dep.var, exog.regressors,
  valid.instruments, instrument1, instrument2, gamma.value)
}
\arguments{
\item{data}{a dataframe that contains all data}

\item{endo.regressor}{the regressor that is correlated with the error term(x)}

\item{dep.var}{the dependend variable (y)}

\item{exog.regressors}{the regressors that are not correlated with dep.var and the error term}

\item{valid.instruments}{instruments that can be considered independent of the error term}

\item{instrument1}{the (imperfect) instrument used to estimate the effect of the endogenous regressor}

\item{instrument2}{the second (imperfect) instrument used to estimate the effect of the endogenous regressor}

\item{gamme.value}{the weight given to the two instruments (0,1). The closed interval would mean one instrument is not used, therefore it does not make sense to used}
}
\value{
the beta bounds calculated using two iiv
}
\description{
This estimator calculates the bounds of the ivv coefficients when more than one imperfect instrument is 
available.
}
\examples{
Two.Instr.Bounds()
}
\keyword{endogenous}
\keyword{iiv,}
\keyword{one}
\keyword{regressor}
\keyword{simple,}
